test(client): extract global mocks in a separate file and fix linting errors
