+++ packages/client/jest.config.js
@@ -4,5 +4,5 @@ module.exports = {
         'node_modules/(?!(jest-)?react-native|react-clone-referenced-element|dripsy|@dripsy/.*|@react-native-community|expo(nent)?|@expo(nent)?/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|@sentry/.*)',
     ],
     testPathIgnorePatterns: ['/node_modules/', 'dist'],
-    setupFilesAfterEnv: ['./jest.setup.ts'],
+    setupFilesAfterEnv: ['./jest.setup.ts', './tests/global-mocks.ts'],
 };
